DDL 변경에 따른 엔티티 수정 및 refresh token을 쿠키에 담아 관리하도록 변경

+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/common/security/AdminResolver.java
@@ -22,7 +22,7 @@ public class AdminResolver implements HandlerMethodArgumentResolver {
 	public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
 								  NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
 		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-		Long adminId = (Long) authentication.getPrincipal();
+		Integer adminId = (Integer) authentication.getPrincipal();
 		AdminRole adminRole = authentication.getAuthorities().stream()
 			.findFirst()
 			.map(authority -> AdminRole.valueOf(authority.getAuthority()))
